Sending necessary request parameters changes

+++ src/main/java/com/ontarget/api/dao/DocumentDAO.java
@@ -9,11 +9,11 @@ public interface DocumentDAO extends GenericDAO<Document>{
 
     public boolean updateStatus(long documentId, String newStatus, int modifiedBy);
 
-    public List<Document> getByCreatedBy(String createdBy, long projectId);
+    public List<Document> getByCreatedBy(String createdBy, int projectId);
 
-    public List<Document> getByAssigneeUsername(String username, long projectId);
+    public List<Document> getByAssigneeUsername(String username, int projectId);
 
-    public List<Document> getDocumentsByProject(long projectId, String approved) throws Exception;
+    public List<Document> getDocumentsByProject(int projectId, String approved) throws Exception;
 
     public boolean updateDueDate(long documentId,Date dueDate, String modifiedBy) throws Exception;
 }
